Refactor extract to a utils library the common code
